+++ hooks/post_gen_project.py
@@ -17,12 +17,24 @@ if __name__ == "__main__":
         except Exception:
             subprocess.check_call([sys.executable, join('ci', 'bootstrap.py')])
 {% endif %}
+{%- if cookiecutter.command_line_interface|lower == 'no' %}
+    from os.path import join
+    import os
+    os.unlink(join('src', '{{ cookiecutter.package_name }}', '__main__.py'))
+{% endif %}
 
     print("""
 ################################################################################
 ################################################################################
 
     You have succesfully created `{{ cookiecutter.repo_name }}`.
+
+################################################################################
+
+    You've used these cookiecutter parameters:
+{% for key, value in cookiecutter.items()|sort %}
+        {{ "%-26s"|format(key + ":") }} {{ (value|pprint).lstrip("u") }}
+{%- endfor %}
 
 ################################################################################