+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ v1.2 (2015-04-10)
   If you run cookiecutter on Python 2 you'll get unicode escapes ("\uXXXX") and on Python 3 you'll get the pretty gliphs.
 * Fixed the ``bootstrap.py`` script (that's used for the ``test_matrix_configurator`` mode) to work from any current working directory.
 * Included the branch name in the AppVeyor build number.
+* Make the CLI optional and add support for using `click`.
 
 v1.1 (2015-03-28)
 -----------------